Latest

+++ router/Addproduct.js
@@ -1,24 +1,28 @@
 const express = require('express');
 const Product = require('../Model/Inventory');
 const shortid = require('shortid');
+const verifyToken = require('../VerifyToken')
 const router = express.Router();
 
 
-router.post('/add-product', async (req, res) => {
+router.post('/add-product', verifyToken ,async (req, res) => {
+    const { Itemcode, Model, price, Brand, Inward, Outward, Current } = req.body;
+
     try {
-         const { Itemcode , Model , price , Brand , Inward , Outward , Current } = req.body;
-         if (!Itemcode || !Model || !price || !Brand || !Inward || !Outward || !Current) {
-             return res.status(400).json({
-                 message: 'Inventory field is required'
-             });
-         }
- 
-         const randomNum = Math.floor(100 + Math.random() * 900);
+        if (req.user.role !== 'Stock Filler') {
+            return res.status(403).json({ message: 'Access denied. You do not have permission to access this page.' });
+        }
+        if (!Itemcode || !Model || !price || !Brand || !Inward || !Outward || !Current) {
+            return res.status(400).json({
+                message: 'Inventory field is required'
+            });
+        }
 
-         const ProductId = `PROD-${randomNum}`;
+        const randomNum = Math.floor(100 + Math.random() * 900);
+        const ProductId = `PROD-${randomNum}`;
 
-         const newProduct = new Product({
-            productId : ProductId,
+        const newProduct = new Product({
+            productId: ProductId,
             Itemcode,
             Model,
             price,
@@ -26,22 +30,24 @@ router.post('/add-product', async (req, res) => {
             Inward,
             Outward,
             Current
-         });
- 
-         const savedProduct = await newProduct.save();
-         return res.status(200).json({
-             message: 'Product added successfully',
-             product: savedProduct
-         });
-     } catch (error) {
-         console.error(error);
-         return res.status(500).json({ message: 'Server error' });
-     }
- });
+        });
+
+        const savedProduct = await newProduct.save();
+        return res.status(200).json({
+            message: 'Product added successfully',
+            product: savedProduct
+        });
+    } catch (error) {
+        console.error(error);
+        return res.status(500).json({ message: 'Server error' });
+    }
+});
+
  
  
  router.get('/get-product',async (req,res) => {
      try {
+        
          const stocks = await Product.find();
          res.status(200).json(stocks)
      } catch (error) {
@@ -54,6 +60,9 @@ router.post('/add-product', async (req, res) => {
  
  router.put('/update/:id',async (req,res) => {
      try {
+        if(req.user && req.user.role !== "Stock Filler"){
+            return res.status(403).json({ message: 'Access denied. You do not have permission to access this page.' });
+        }
          const updatedstock = await Product.findByIdAndUpdate(req.params.id , req.body)
          res.status(200).json(updatedstock)
      } catch (error) {
@@ -66,6 +75,9 @@ router.post('/add-product', async (req, res) => {
  
  router.delete('/delete/:id', async (req,res) => {
      try {
+        if(req.user && req.user.role !== "Stock Filler"){
+            return res.status(403).json({ message: 'Access denied. You do not have permission to access this page.' });
+        }
          const deleteproduct = await Product.findByIdAndDelete(req.params.id)
          res.status(200).json(deleteproduct)
      } catch (error) {
